+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added vibrato, tremolo, more accurate portamento support, and fixed pan calculation in the CapcomSnes driver.
 - Added drum kit support to the SuzukiSnes driver (Seiken Densetsu 3, Super Mario RPG, etc)
 - Improved support for late-era Intelligent Systems SNES/SFC titles (Tetris Attack, Fire Emblem 4, Super Famicom Wars) 
+- Added support for the missing GBA (Sappy aka mp2k) commands (LFO, pitch bend range, transpose, ...)
 
 ### Changed